requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.14.0
fpdf2>=2.8.7
python-pptx==1.0.2
